Move match exactly sync setup to BackgroundWorkManager
