+++ collect_app/src/main/java/org/odk/collect/android/backgroundwork/BackgroundWorkManager.java
@@ -1,0 +1,10 @@
+package org.odk.collect.android.backgroundwork;
+
+public interface BackgroundWorkManager {
+
+    boolean isFormUploaderRunning();
+
+    boolean isFormDownloaderRunning();
+
+    void scheduleMatchExactlySync();
+}
