--- collect_app/src/main/java/org/odk/collect/utilities/BackgroundWorkManager.java
@@ -1,6 +1,0 @@
-package org.odk.collect.utilities;
-
-public interface BackgroundWorkManager {
-
-    boolean isRunning(String tag);
-}
